requirements.txt
fastapi
uvicorn
jinja2
pydantic
python-multipart
asyncio
paho-mqtt==1.6.1
opcua
cryptography
influxdb
influxdb-client
asyncua